修正

+++ todo/views.py
@@ -1,50 +1,63 @@
-from django.shortcuts import render,redirect
 from django.shortcuts import render, redirect
 from django.http import Http404
 from django.utils.timezone import make_aware
 from django.utils.dateparse import parse_datetime
-from todo.models import Task 
-# Create your views here.
+from todo.models import Task
+
 def index(request):
     if request.method == 'POST':
-        task= Task(title=request.POST['title'],
-                   due_at=make_aware (parse_datetime (request.POST['due_at'])))
+        task = Task(
+            title=request.POST['title'],
+            due_at=make_aware(parse_datetime(request.POST['due_at']))
+        )
         task.save()
     if request.GET.get('order') == 'due':
         tasks = Task.objects.order_by('due_at')
     else:
         tasks = Task.objects.order_by('-posted_at')
-    context = { 
+    context = {
         'tasks': tasks
     }
-    return render (request, 'todo/index.html', context)
+    return render(request, 'todo/index.html', context)
 
 def detail(request, task_id):
     try:
         task = Task.objects.get(pk=task_id)
     except Task.DoesNotExist:
         raise Http404("Task does not exist")
-    task.delete()
+    context = {
+        'task': task,
+    }
     return render(request, 'todo/detail.html', context)
 
-
 def update(request, task_id):
-def delete(request, task_id):
     try:
         task = Task.objects.get(pk=task_id)
     except Task.DoesNotExist:
         raise Http404("Task does not exist")
-    
+
     if request.method == 'POST':
         task.title = request.POST['title']
         task.due_at = make_aware(parse_datetime(request.POST['due_at']))
         task.save()
         return redirect('detail', task_id=task.id)
-    
+
     context = {
         'task': task,
     }
     return render(request, 'todo/edit.html', context)
 
-    task.delete()
-    return redirect(index)
+def delete(request, task_id):
+    try:
+        task = Task.objects.get(pk=task_id)
+    except Task.DoesNotExist:
+        raise Http404("Task does not exist")
+
+    if request.method == 'POST':
+        task.delete()
+        return redirect('index')
+
+    context = {
+        'task': task,
+    }
+    return render(request, 'todo/delete.html', context)
